requirements.txt
factor_analyzer==0.5.1
gql==3.5.2
ipython==8.12.3
joblib==1.4.2
matplotlib==3.7.5
numpy==1.26.4
pandas==2.0.3
pg8000==1.31.2
psycopg2_binary==2.9.10
pygam==0.9.1
python-dotenv==1.0.1
requests==2.32.3
requests_toolbelt==1.0.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
tqdm==4.67.1
xgboost==2.1.3
shapely==2.0.7
penaltyblog==1.4.0
pyarrow==20.0.0
sparse==0.17.0
statsmodels==0.14.0